install: introduce symlink cache in LookupPaths
This is an optimization aimed to speedup ListUnitFiles() on systems with
a lot of units. Because unit file enablement state is given by the state
of filesystem we need to interrogate all directories of
interest (e.g. .wants) and all symlinks in them to see whether some of
them actually points to unit file we are examining. While going through
them we encounter a lot of symlinks that point to different
unit file. Let's not throw away information about them , instead put it
in a cache and use it in a future to save context swithcing in kernel to
do readlinkat() system calls.

Before (on Fedora 33, with additional 1000 dummy service unit files),

$ strace -e readlinkat -c systemctl --no-legend --no-pager --root=/ list-unit-files >/dev/null
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
100.00    2.202122           1   1239556           readlinkat
------ ----------- ----------- --------- --------- ----------------
100.00    2.202122           1   1239556           total

$ time systemctl --no-legend --no-pager --root=/ list-unit-files >/dev/null

real    0m3.236s
user    0m1.207s
sys     0m1.886s

After,

$ strace -e readlinkat -c ./systemctl --no-legend --no-pager --root=/ list-unit-files >/dev/null
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ----------------
100.00    0.001806           1      1235           readlinkat
------ ----------- ----------- --------- --------- ----------------
100.00    0.001806           1      1235           total

$ time ./systemctl --no-legend --no-pager --root=/ list-unit-files >/dev/null

real    0m2.580s
user    0m1.689s
sys     0m0.770s

Speed up overall isn't massive and we are trading system time for user time
but still it seems worth doing.