core: Add unit name in unit_status_string()
The core of systemd is very oriented around unit names; I
really love things like `systemctl status foo`.

But the problem is that the "default" systemd output for
both the journal and the bootup status print the unit *description*
and don't include the unit name.

Now, for a lot of built-in systemd units I "know" the unit name
or it's pretty obvious.  When one starts talking about 3rd party
units, some of the descriptions are quite different from the
actual unit name.  For example, crio was only recently improved:
https://github.com/cri-o/cri-o/commit/564d5f0b22e666a3ecf703cf162c2e233ccd297c

I so often want to go from looking at e.g. `journalctl` (or the serial console)
outputting a failure to log in and interactively debug, and even
for units I'm familiar with I sometimes still have to crudely do e.g.
`systemctl list-units | grep -e 'Some Description'` to find the unit name.

This patch functions, but it's a bit more of an RFC because if
we really decided to do this properly there's a lot of other cleanup
that needs to happen.  For example
`May 28 22:36:13 cosa-devsh systemd[1]: Reached target basic.target: Basic System.`
is obviously redundant; probably should be more like:
`May 28 22:36:13 cosa-devsh systemd[1]: Reached basic.target: Basic System.`