networkd: re-assign prefix after DHCPv6 PD update
Previously on a lease update/renewal prefixes would not be announced by
radv anymore. Since PD prefixes can have a very long lifetime
(days/weeks/…) that issue wasn't really obvious in the previous (manual)
tests.

Instead of bailing out whenever there is already a PD prefix configured
on the interface we now let the code continue to re-assign the prefix as
long as it is the same prefix (as before).